+++ lib/graphql/execution/deferred_execution.rb
@@ -165,74 +165,6 @@ module GraphQL
         initial_result
       end
 
-      # Global, immutable environment for executing `query`.
-      # Passed through all execution to provide type, fragment and field lookup.
-      class ExecScope
-        attr_reader :query, :schema
-
-        def initialize(query)
-          @query = query
-          @schema = query.schema
-        end
-
-        def get_type(type)
-          @schema.types[type]
-        end
-
-        def get_fragment(name)
-          @query.fragments[name]
-        end
-
-        # This includes dynamic fields like __typename
-        def get_field(type, name)
-          @schema.get_field(type, name) || raise("No field named '#{name}' found for #{type}")
-        end
-      end
-
-      # One serial stream of execution. One thread runs the initial query,
-      # then any deferred frames are restarted with their own threads.
-      #
-      # - {ExecThread#errors} contains errors during this part of the query
-      # - {ExecThread#defers} contains {ExecFrame}s which were marked as `@defer`
-      #   and will be executed with their own threads later.
-      class ExecThread
-        attr_reader :errors, :defers
-        def initialize
-          @errors = []
-          @defers = []
-        end
-      end
-
-      # One step of execution. Each step in execution gets its own frame.
-      #
-      # - {ExecFrame#node} is the IRep node which is being interpreted
-      # - {ExecFrame#path} is like a stack trace, it is used for patching deferred values
-      # - {ExecFrame#value} is the object being exposed by GraphQL at this point
-      # - {ExecFrame#type} is the GraphQL type which exposes {#value} at this point
-      class ExecFrame
-        attr_reader :node, :path, :type, :value
-        def initialize(node:, path:, type:, value:)
-          @node = node
-          @path = path
-          @type = type
-          @value = value
-        end
-      end
-
-      # Contains the list field's ExecFrame
-      # And the enumerator which is being mapped
-      # - {ExecStream#enumerator} is an Enumerator which yields `item, idx`
-      # - {ExecStream#frame} is the {ExecFrame} for the list selection (where `@stream` was present)
-      # - {ExecStream#type} is the inner type of the list (the item's type)
-      class ExecStream
-        attr_reader :enumerator, :frame, :type
-        def initialize(enumerator:, frame:, type:)
-          @enumerator = enumerator
-          @frame = frame
-          @type = type
-        end
-      end
-
       private
 
       # If this `frame` is marked as defer, add it to `defers`
